Refactor annotations component in entity tooltip.
